execution/bal: extract more bal code out of stagedsync pkg
