+++ execution/stagedsync/bal_load_test.go
@@ -18,7 +18,6 @@ package stagedsync
 
 import (
 	"bytes"
-	"math"
 	"testing"
 
 	"github.com/holiman/uint256"
@@ -31,67 +30,6 @@ import (
 	"github.com/erigontech/erigon/execution/types"
 	"github.com/erigontech/erigon/execution/types/accounts"
 )
-
-// finalChange returns the last (block-end) change; BAL change lists reach it
-// already validated strictly-increasing by tx Index.
-func TestFinalChange(t *testing.T) {
-	t.Parallel()
-
-	if _, ok := finalChangeUpTo([]*types.BalanceChange(nil), math.MaxUint32); ok {
-		t.Fatal("finalChangeUpTo on an empty list returned ok=true")
-	}
-
-	single, ok := finalChangeUpTo([]*types.BalanceChange{
-		{Index: 7, Value: *uint256.NewInt(70)},
-	}, math.MaxUint32)
-	require.True(t, ok)
-	assert.Equal(t, uint64(70), single.Value.Uint64())
-
-	ascending, ok := finalChangeUpTo([]*types.BalanceChange{
-		{Index: 0, Value: *uint256.NewInt(10)},
-		{Index: 2, Value: *uint256.NewInt(20)},
-		{Index: 9, Value: *uint256.NewInt(90)},
-	}, math.MaxUint32)
-	require.True(t, ok)
-	assert.Equal(t, uint64(90), ascending.Value.Uint64())
-}
-
-// finalChangeUpTo returns the field's value as of a mid-block tx index — the
-// primitive that lets the fold checkpoint at a step boundary from the per-tx BAL.
-func TestFinalChangeUpTo(t *testing.T) {
-	t.Parallel()
-
-	changes := []*types.BalanceChange{
-		{Index: 0, Value: *uint256.NewInt(10)},
-		{Index: 2, Value: *uint256.NewInt(20)},
-		{Index: 9, Value: *uint256.NewInt(90)},
-	}
-
-	if _, ok := finalChangeUpTo([]*types.BalanceChange{{Index: 5, Value: *uint256.NewInt(1)}}, 4); ok {
-		t.Fatal("ceiling below the only change must return ok=false")
-	}
-
-	// Exactly at, and between, change indices → latest change ≤ ceiling.
-	at2, ok := finalChangeUpTo(changes, 2)
-	require.True(t, ok)
-	assert.Equal(t, uint64(20), at2.Value.Uint64(), "ceiling==Index returns that change")
-
-	between, ok := finalChangeUpTo(changes, 5)
-	require.True(t, ok)
-	assert.Equal(t, uint64(20), between.Value.Uint64(), "ceiling between changes returns the earlier one")
-
-	// At/above the last index → whole-block value (== finalChange).
-	whole, ok := finalChangeUpTo(changes, 9)
-	require.True(t, ok)
-	assert.Equal(t, uint64(90), whole.Value.Uint64())
-	all, ok := finalChangeUpTo(changes, ^uint32(0))
-	require.True(t, ok)
-	assert.Equal(t, uint64(90), all.Value.Uint64(), "MaxUint32 ceiling == finalChange")
-
-	if _, ok := finalChangeUpTo([]*types.BalanceChange(nil), 100); ok {
-		t.Fatal("finalChangeUpTo on an empty list returned ok=true")
-	}
-}
 
 // TestLoadFromBAL_MatchesApplyWrites is a differential test:
 // loading calcState from a BAL must produce exactly the same accumulated
